requirements.txt
zai-sdk>=0.0.3.3